Rename files in js/src/settings/reconnect-accounts
